fix(hub-common): move function to add default predicate to util
